adminlte3/requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2.10
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==20.0.4
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
msgpack==1.0.8
pillow==10.3.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
psycopg2-binary==2.9.9
pyOpenSSL==24.1.0
PyYAML==6.0.1
redis==5.0.4
referencing==0.34.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing-extensions==4.11.0
uritemplate==4.1.1
zope.interface==6.3